usa011-Ingredients-crud #11
Changement du code pour faire des TypeQuerys:
- modification de IngredientService.

+++ pizzeria-admin-web/src/main/java/fr/pizzeria/admin/metier/IngredientService.java
@@ -3,40 +3,67 @@ package fr.pizzeria.admin.metier;
 import java.util.List;
 
 import javax.ejb.Stateless;
+import javax.ejb.TransactionAttribute;
+import javax.ejb.TransactionAttributeType;
+import javax.ejb.TransactionManagement;
+import javax.ejb.TransactionManagementType;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
+import javax.persistence.TypedQuery;
 
 import fr.pizzeria.model.Ingredient;
 
 @Stateless
+@TransactionManagement(value = TransactionManagementType.CONTAINER)
 public class IngredientService {
 	
-	@PersistenceContext
+	@PersistenceContext(unitName = "pizzeria-admin-web")
 	private EntityManager em;
 	
+	private TypedQuery<Ingredient> query;
+	
+	@TransactionAttribute(TransactionAttributeType.REQUIRED)
 	public List<Ingredient> findAll(){
-		return em.createQuery("select i from Ingredient i", Ingredient.class).getResultList();
+		this.query = em.createQuery("select i from Ingredient i", Ingredient.class);
+		List<Ingredient> ingredients = query.getResultList();
+		//ingredients != null ? return ingredients : //todo exception
+		return ingredients;
 	}
 	
+	@TransactionAttribute(TransactionAttributeType.REQUIRED)
 	public Ingredient findById(Integer id){
-		return em.createQuery("select i from Ingredient i where i.id=:id", Ingredient.class)
-				.setParameter("id", id).getSingleResult();
+		
+		this.query = em.createQuery("select i from Ingredient i where i.id=:id", Ingredient.class);
+		this.query.setParameter("id", id);
+		
+		return query.getSingleResult();
 	}
 	
+	@TransactionAttribute(TransactionAttributeType.REQUIRED)
 	public void save(Ingredient ingredient) {
 		em.persist(ingredient);
 	}
 	
+	@TransactionAttribute(TransactionAttributeType.REQUIRED)
 	public void update(Integer id, Ingredient ingredient){
+		
 		Ingredient old = findById(id);
 		
-		ingredient.setId(old.getId());
-		em.merge(ingredient);
+		if(old != null){
+			ingredient.setId(old.getId());
+			em.merge(ingredient);			
+		}
+		//todo else exception
+		
 	}
 	
+	@TransactionAttribute(TransactionAttributeType.REQUIRED)
 	public void delete(Integer id){
 		Ingredient ingredient = findById(id);
-		em.remove(ingredient);
+		if(ingredient != null){
+			em.remove(ingredient);			
+		}
+		//todo else exception
 	}
 	
 }
